fix: Fix locations not being saved (resetting upon restart)
This was potentially caused becasue by using the raw player's name for a
save game, for maintaining uniqueness across multiple save games.
The error occured when the player's name contains space or any character
other than letters, numbers, underscores, hyphens or dots.
And also if the player's name is same across two or more saves, this
would result in both of them sharing the same travel locations.

This update will now use the player's name (with only the allowed chars)
and the unique id for the save game, fixing both the issues.